% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlspice_package.R
\docType{package}
\name{emlspice}
\alias{emlspice}
\alias{emlspice-package}
\title{\code{emlspice} package}
\description{
Converts EML to Tabular Formats
}
\details{
See \href{https://github.com/isteves/emlspice}{GitHub}
}
