% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_access.R
\name{es_access}
\alias{es_access}
\title{Get access from EML}
\usage{
es_access(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) file path for saving the table to disk}
}
\description{
Return EML access in the dataspice access.csv format.
}
\examples{
\dontrun{
eml_path <- system.file("LeConte_meteo_metadata.xml", package = "emlspice")
eml <- read_eml(eml_path)
es_access(eml)
}
}
