% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_attributes.R
\name{es_attributes}
\alias{es_attributes}
\title{Get attributes from EML}
\usage{
es_attributes(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) file path for saving the table to disk}
}
\description{
Return EML attributes in the dataspice attributes.csv format.
}
\examples{
\dontrun{
eml_path <- system.file("LeConte_meteo_metadata.xml", package = "emlspice")
eml <- read_eml(eml_path)
es_attributes(eml)
}
}
