% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_biblio.R
\name{es_biblio}
\alias{es_biblio}
\title{Get biblio from EML}
\usage{
es_biblio(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) file path for saving the table to disk}
}
\description{
Return EML biblio in the dataspice biblio.csv format.
}
\examples{
\dontrun{
eml_path <- system.file("LeConte_meteo_metadata.xml", package = "emlspice")
eml <- read_eml(eml_path)
es_biblio(eml)
}
}
