% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_creators.R
\name{es_creators}
\alias{es_creators}
\title{Get creators from EML}
\usage{
es_creators(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) file path for saving the table to disk}
}
\description{
Return EML creators in the dataspice creators.csv format.
}
\examples{
\dontrun{
eml_path <- system.file("LeConte_meteo_metadata.xml", package = "emlspice")
eml <- read_eml(eml_path)
es_creators(eml)
}
}
