% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emlspice.R
\name{get_emlspice}
\alias{get_emlspice}
\title{Get dataspice tabular formats from EML}
\usage{
get_emlspice(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) folder path for saving the table to disk}
}
\description{
Return EML in the dataspice dataframes.
}
\examples{
\dontrun{
eml_path <- system.file("LeConte_meteo_metadata.xml", package = "emlspice")
eml <- read_eml(eml_path)
get_emlspice(eml)
}
}
